Refactor serialization
